containers/query_manager/requirements.txt
couchbase==4.2.1
dnspython
requests